Update readme to fix tests.
The tests as previously written in the readme did not work. When you check state immediately after setting it, the value has not changed yet. A `setTimeout` solves the problem, but I used `tickedoff` in the example instead, since that is what `unstated` itself is using.

If there is a better way to do this, I would love to see it. The nested `defer` functions can get messy. I tried using async/await, but `tickedoff` doesn't support that. Also tried using `jest.useFakeTimers()` and `jest.runAllTimers()` but since `unstated` does not use timeouts, this didn't work at all.