Information on contributing and issues
I'm creating this because there is no way to report issues in Unstated and there is no guide on how to contribute. Is there a reason for this? It concerns me that there is none of this information available. There is also no license in the project root, which can be an issue for some people.

Is there a place to report bugs? Request features? Learn how to contribute?